更换了评论UI，评论以及回复功能完成

+++ zeus-manager/zeus-mapper/src/main/java/com/nzl/dao/UserMapper.java
@@ -43,6 +43,6 @@ public interface UserMapper extends Mapper<UserDto> {
      * @param replyId
      * @return
      */
-    String getUsernameByReplyId(Integer replyId);
+    String getUsernameByReplyId(Long replyId);
 
 }
